Player adaptation update: fixed bugs, screen interactions with animation
