+++ app/build.gradle
@@ -5,6 +5,7 @@ plugins {
     id 'org.jetbrains.kotlin.plugin.serialization' version '1.7.10'
     id("com.google.dagger.hilt.android")
     id 'kotlin-kapt'
+    id 'kotlin-parcelize'
 }
 
 android {
