+++ app/src/main/java/com/example/musicapp/data/MusicRepository.kt
@@ -1,6 +1,5 @@
 package com.example.musicapp.data
 
-import android.app.Application
 import androidx.media3.exoplayer.ExoPlayer
 
 interface MusicRepository {
