+++ app/src/main/java/com/example/musicapp/model/Artist.kt
@@ -1,23 +1,8 @@
 package com.example.musicapp.model
 
+import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
-
-/*@kotlinx.serialization.Serializable
-data class Artist(
-    val external_urls: ExternalUrls,
-    val href: String,
-    val id: String,
-    val name: String,
-    val type: String,
-    val uri: String
-)
-
-@kotlinx.serialization.Serializable
-data class ExternalUrls(
-    val spotify: String
-)
-*/
-
+import kotlinx.parcelize.Parcelize
 
 @kotlinx.serialization.Serializable
 data class Artist(
