+++ app/src/main/java/com/example/musicapp/ui/mainScreen/MainScreen.kt
@@ -1,16 +1,12 @@
 package com.example.musicapp.ui.mainScreen
 
-import MusicViewModel
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.pullrefresh.PullRefreshState
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavController
 import com.example.musicapp.network.NetworkUiState
 import com.example.musicapp.network.NetworkViewModel
 
 
-//@Preview(showSystemUi = true)
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun MainScreen(
