+++ app/src/main/java/com/example/musicapp/ui/musicScreen/PlayerButtons.kt
@@ -1,15 +1,21 @@
 package com.example.musicapp.ui.musicScreen
 
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.indication
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
+import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.semantics.Role
@@ -21,27 +27,38 @@ import com.example.musicapp.R
 
 @Composable
 fun PlayerButtons(
+    onPlayPauseClicked: () -> Unit,
+    onNextClicked: () -> Unit,
+    onPrevClicked: () -> Unit,
+    onRewindClicked: () -> Unit,
+    onForwardClicked: () -> Unit,
+    playIcon: ImageVector,
     playerButtonSize: Dp = 72.dp,
     sideButtonSize: Dp = 42.dp,
     modifier: Modifier = Modifier,
 ) {
 
-    
+
     Row(
         modifier = modifier.fillMaxWidth(),
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.SpaceEvenly
     ) {
-        val buttonModifier = Modifier
-            .size(sideButtonSize)
-            .semantics { role = Role.Button }
+
 
         Image(
             painter = painterResource(id = R.drawable.skip_previous),
             contentDescription = "Skip Previous",
             contentScale = ContentScale.Fit,
             colorFilter = ColorFilter.tint(Color.Black),
-            modifier = buttonModifier
+            modifier = Modifier
+                .size(sideButtonSize)
+                .clickable(
+                    indication = rememberRipple(bounded = false),
+                    interactionSource = remember { MutableInteractionSource() },
+                    onClick = onPrevClicked,
+                    role = Role.Button
+                )
         )
 
         Image(
@@ -49,17 +66,29 @@ fun PlayerButtons(
             contentDescription = "Reply 10 Second",
             contentScale = ContentScale.Fit,
             colorFilter = ColorFilter.tint(Color.Black),
-            modifier = buttonModifier
+            modifier = Modifier
+                .size(sideButtonSize)
+                .clickable(
+                    indication = rememberRipple(bounded = false),
+                    interactionSource = remember { MutableInteractionSource() },
+                    onClick = onRewindClicked,
+                    role = Role.Button
+                )
         )
 
         Image(
-            painter = painterResource(id = R.drawable.play_circle),
-            contentDescription = "Play",
+            imageVector = playIcon,
+            contentDescription = "Play/Pause",
             contentScale = ContentScale.Fit,
             colorFilter = ColorFilter.tint(Color.Black),
             modifier = Modifier
                 .size(playerButtonSize)
-                .semantics { role = Role.Button }
+                .clickable(
+                    indication = rememberRipple(bounded = false),
+                    interactionSource = remember { MutableInteractionSource() },
+                    onClick = onPlayPauseClicked,
+                    role = Role.Button
+                )
         )
 
         Image(
@@ -67,7 +96,14 @@ fun PlayerButtons(
             contentDescription = "Forward 10 Seconds",
             contentScale = ContentScale.Fit,
             colorFilter = ColorFilter.tint(Color.Black),
-            modifier = buttonModifier
+            modifier = Modifier
+                .size(sideButtonSize)
+                .clickable(
+                    indication = rememberRipple(bounded = false),
+                    interactionSource = remember { MutableInteractionSource() },
+                    onClick = onForwardClicked,
+                    role = Role.Button
+                )
         )
 
         Image(
@@ -75,7 +111,14 @@ fun PlayerButtons(
             contentDescription = "Skip Next",
             contentScale = ContentScale.Fit,
             colorFilter = ColorFilter.tint(Color.Black),
-            modifier = buttonModifier
+            modifier = Modifier
+                .size(sideButtonSize)
+                .clickable(
+                    indication = rememberRipple(bounded = false),
+                    interactionSource = remember { MutableInteractionSource() },
+                    onClick = onNextClicked,
+                    role = Role.Button
+                )
         )
     }
 }
