+++ app/src/main/java/com/example/musicapp/ui/musicScreen/constants.kt
@@ -1,0 +1,3 @@
+package com.example.musicapp.ui.musicScreen
+
+const val FORWARD_BACKWARD_STEP = 5000
