+++ app/src/main/java/com/example/musicapp/ui/navigation/graphs/HomeNavGraph.kt
@@ -1,29 +1,18 @@
 package com.example.musicapp.ui.navigation.graphs
 
-import MusicViewModel
-import android.util.Log
-import androidx.compose.foundation.layout.Box
+import com.example.musicapp.viewmodel.MusicViewModel
 import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.Text
 import androidx.compose.material.pullrefresh.rememberPullRefreshState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.platform.LocalContext
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.*
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import androidx.navigation.compose.navigation
-import com.example.musicapp.data.BottomNavItems
-import com.example.musicapp.data.MusicRepositoryManager
 import com.example.musicapp.network.NetworkUiState
 import com.example.musicapp.network.NetworkViewModel
-import com.example.musicapp.ui.AlbumScreen.AlbumScreen
-import com.example.musicapp.ui.mainScreen.HomeScreen
+import com.example.musicapp.ui.albumScreen.AlbumScreen
 import com.example.musicapp.ui.mainScreen.MainScreen
-import com.example.musicapp.ui.musicScreen.MusicScreen
 import com.example.musicapp.ui.navigation.NavRoutes
 
 
@@ -76,29 +65,28 @@ fun HomeNavGraph(
 
             musicViewModel.updatePlaylist(NetworkUiState.trackResponse!!)
             AlbumScreen(
-                musicViewModel,
-                NetworkUiState.playlistsResponse!![albumId],
-                NetworkUiState.trackResponse!!,
-                onClick = { trackIndex ->
-                    navController.navigate(NavRoutes.Track.name + "/$trackIndex")
-                }
+                musicViewModel = musicViewModel,
+                playlist = NetworkUiState.playlistsResponse!![albumId],
+                tracks = NetworkUiState.trackResponse!!,
             )
         }
-        composable(
-            route = NavRoutes.Track.name + "/{track}",
-            arguments = listOf(navArgument("track") {
-                type = NavType.IntType
-            })
+        /*composable(
+            route = NavRoutes.Track.name + "/{state}",
+            arguments = listOf(
+                navArgument("track") { type = NavType.IntType },
+               navArgument("state") { type = StateType() }
+
+            )
         ) { navBackStackEntry ->
             val trackId =
                 navBackStackEntry.arguments?.getInt("track") ?: throw NullPointerException()
-
             MusicScreen(
-                //track = NetworkUiState.trackResponse!!.items[trackId],
-                track = NetworkUiState.trackResponse!![trackId],
+                track = NetworkUiState.trackResponse!![state.currentTrackIndex],
+                musicViewModel = musicViewModel,
+                musicUiState = state,
                 onCollapseClicked = { navController.popBackStack() }
             )
-        }
+        }*/
     }
 }
