+++ app/src/main/java/com/example/musicapp/viewmodel/MusicViewModel.kt
@@ -1,27 +1,27 @@
+package com.example.musicapp.viewmodel
+
 import android.net.Uri
-import android.util.Log
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
+import android.os.Handler
+import android.os.Looper
+import android.os.Parcelable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import androidx.media3.common.AudioAttributes
 import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
 import androidx.media3.exoplayer.ExoPlayer
 import com.example.musicapp.MusicApplication
 import com.example.musicapp.data.MusicRepository
-import com.example.musicapp.model.Playlist
-import com.example.musicapp.model.PlaylistTracksResponse
 import com.example.musicapp.model.Track
 import kotlinx.coroutines.flow.*
+import kotlinx.parcelize.Parcelize
 
 data class MusicUiState(
     val isPlaying: Boolean = false,
     val playlist: List<Track> = emptyList(),
     val currentTrackIndex: Int = 0,
+    val currentPosition: Long = 0,
 )
 
 class MusicViewModel(private val musicRepository: MusicRepository) : ViewModel() {
@@ -29,9 +29,9 @@ class MusicViewModel(private val musicRepository: MusicRepository) : ViewModel()
     private val _musicUiState = MutableStateFlow(MusicUiState())
     val musicUiState: StateFlow<MusicUiState> = _musicUiState.asStateFlow()
 
+    val player: ExoPlayer = musicRepository.getExoPlayerInstance()
 
-    var player: ExoPlayer = musicRepository.getExoPlayerInstance()
-
+    private val _handler = Handler(Looper.getMainLooper())
     init {
         player.repeatMode = ExoPlayer.REPEAT_MODE_ALL
 
@@ -46,21 +46,32 @@ class MusicViewModel(private val musicRepository: MusicRepository) : ViewModel()
         )
     }
 
-
     fun updatePlaylist(playlist: List<Track>) {
 
         player.clearMediaItems()
         playlist.forEach {
             player.addMediaItem(MediaItem.fromUri(Uri.parse(it.preview)))
         }
-        _musicUiState.update { state -> state.copy(isPlaying = false, playlist = playlist) }
+        _musicUiState.update { state -> state.copy(isPlaying = false, playlist = playlist, currentPosition = 0L) }
         player.prepare()
         player.seekTo(0, 0L)
     }
 
     fun play() {
-        _musicUiState.update { it.copy(isPlaying = true) }
+        _musicUiState.update { it.copy(isPlaying = true, currentTrackIndex = player.currentMediaItemIndex) }
         player.play()
+
+        _handler.postDelayed(object: Runnable {
+            override fun run() {
+                try {
+                    _musicUiState.update { it.copy(currentPosition = player.currentPosition) }
+                    _handler.postDelayed(this, 0)
+                }
+                catch (ex: Exception) {
+                    _musicUiState.update { it.copy(currentPosition = 0L) }
+                }
+            }
+        }, 0)
     }
     fun pause() {
         _musicUiState.update { it.copy(isPlaying = false) }
@@ -80,35 +91,45 @@ class MusicViewModel(private val musicRepository: MusicRepository) : ViewModel()
         _musicUiState.update { it.copy(isPlaying = true, currentTrackIndex = index) }
         player.seekTo(index, 0L)
         player.play()
+
+        _handler.postDelayed(object: Runnable {
+            override fun run() {
+                try {
+                    _musicUiState.update { it.copy(currentPosition = player.currentPosition) }
+                    _handler.postDelayed(this, 0)
+                }
+                catch (ex: Exception) {
+                    _musicUiState.update { it.copy(currentPosition = 0L) }
+                }
+            }
+        }, 0)
     }
 
     private fun stop() {
         player.stop()
         player.release()
-        _musicUiState.update { it.copy(isPlaying = false) }
+        _musicUiState.update { it.copy(isPlaying = false, currentPosition = 0L) }
     }
     fun playNextTrack() {
-
         player.seekToNextMediaItem()
-
         _musicUiState.update { state -> state.copy(isPlaying = true, currentTrackIndex = player.currentMediaItemIndex) }
-        Log.i("PLAYER", "PLAY NEXT AUDIO")
         player.play()
-
     }
 
-    private fun seek(position: Long) {
-        player.seekTo(position)
-    }
+
     fun playPreviousTrack() {
-
         player.seekToPreviousMediaItem()
         _musicUiState.update { state -> state.copy(isPlaying = true, currentTrackIndex = player.currentMediaItemIndex) }
-        Log.i("PLAYER", "PLAY PREViOUS AUDIO")
         player.play()
     }
 
-
+    fun seek(position: Long) {
+        player.seekTo(position)
+        _musicUiState.update {
+            state -> state.copy(isPlaying = true, currentPosition = position)
+        }
+        player.play()
+    }
     companion object {
         val Factory: ViewModelProvider.Factory = viewModelFactory {
             initializer {
